# ods_api_server/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: HPP-V2
  contact: {}
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/Marquardt-Informatik/HPP-V2/1.0.0
  description: SwaggerHub API Auto Mocking
- url: https://virtserver.swaggerhub.com/Marquardt-Informatik/HPP/1.0.0.0
  description: SwaggerHub API Auto Mocking
- url: http://p-nav-as-1-m.int.hrs.com:7047/DynNavHRS/ws/HRS/Codeunit/HPP?WSDL
...
paths:
  /ChangeInvoiceAddress:
    post:
      tags:
      - HPP_Binding
      summary: ChangeInvoiceAddress
      operationId: change_invoice_address
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeInvoiceAddress_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ChangeCustomerAddress:
    post:
      tags:
      - HPP_Binding
      summary: ChangeCustomerAddress
      operationId: change_customer_address
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeCustomerAddress_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_1'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ChangeContactInformation:
    post:
      tags:
      - HPP_Binding
      summary: ChangeContactInformation
      operationId: change_contact_information
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeContactInformation_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /InsertPayment:
    post:
      tags:
      - HPP_Binding
      summary: InsertPayment
      operationId: insert_payment
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertPayment_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_3'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ModifyInvoice:
    post:
      tags:
      - HPP_Binding
      summary: ModifyInvoice
      operationId: modify_invoice
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyInvoice_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_4'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /InsertPaymentV2:
    post:
      tags:
      - HPP_Binding
      summary: InsertPaymentV2
      operationId: insert_payment_v2
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertPaymentV2_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_5'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ChangeCustomerAddressV2:
    post:
      tags:
      - HPP_Binding
      summary: ChangeCustomerAddressV2
      operationId: change_customer_address_v2
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeCustomerAddressV2_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_6'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ModifyInvoiceV2:
    post:
      tags:
      - HPP_Binding
      summary: ModifyInvoiceV2
      operationId: modify_invoice_v2
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyInvoiceV2_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_7'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ModifyInvoiceV3:
    post:
      tags:
      - HPP_Binding
      summary: ModifyInvoiceV3
      operationId: modify_invoice_v3
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyInvoiceV3_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_8'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /InsertPaymentV3:
    post:
      tags:
      - HPP_Binding
      summary: InsertPaymentV3
      operationId: insert_payment_v3
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertPaymentV3_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_9'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ModifyInvoiceV4:
    post:
      tags:
      - HPP_Binding
      summary: ModifyInvoiceV4
      operationId: modify_invoice_v4
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyInvoiceV4_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_10'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ConfirmInvoice:
    post:
      tags:
      - HPP_Binding
      summary: ConfirmInvoice
      operationId: confirm_invoice
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmInvoice_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_11'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ModifyHotelCreditCardInfo:
    post:
      tags:
      - HPP_Binding
      summary: ModifyHotelCreditCardInfo
      operationId: modify_hotel_credit_card_info
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyHotelCreditCardInfo_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_12'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /ModifyHotelBankAccountInfo:
    post:
      tags:
      - HPP_Binding
      summary: ModifyHotelBankAccountInfo
      operationId: modify_hotel_bank_account_info
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModifyHotelBankAccountInfo_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_13'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
  /InsertPaymentPC:
    post:
      tags:
      - HPP_Binding
      summary: InsertPaymentPC
      operationId: insert_payment_pc
      parameters: []
      requestBody:
        description: ""
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InsertPaymentPC_body'
        required: true
      responses:
        "200":
          description: Successfully retrieved the response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_14'
      deprecated: false
      x-openapi-router-controller: ods_api_server.controllers.hpp_binding_controller
components:
  schemas:
...
    ChangeContactInformation:
      title: ChangeContactInformation
      required:
      - ChangeContactInformation
      type: object
      properties:
        ChangeContactInformation:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeContactInformation'
          - {}
    ChangeContactInformation_Result:
      title: ChangeContactInformation_Result
      required:
      - ChangeContactInformation_Result
      type: object
      properties:
        ChangeContactInformation_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeContactInformation_Result'
          - xml:
              name: ChangeContactInformation_Result
              attribute: false
              wrapped: false
      xml:
        name: ChangeContactInformation_Result
        attribute: false
        wrapped: false
    ChangeCustomerAddress:
      title: ChangeCustomerAddress
      required:
      - ChangeCustomerAddress
      type: object
      properties:
        ChangeCustomerAddress:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeCustomerAddress'
          - {}
    ChangeCustomerAddress_Result:
      title: ChangeCustomerAddress_Result
      required:
      - ChangeCustomerAddress_Result
      type: object
      properties:
        ChangeCustomerAddress_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeCustomerAddress_Result'
          - xml:
              name: ChangeCustomerAddress_Result
              attribute: false
              wrapped: false
      xml:
        name: ChangeCustomerAddress_Result
        attribute: false
        wrapped: false
    ChangeCustomerAddressV2_Result:
      title: ChangeCustomerAddressV2_Result
      required:
      - ChangeCustomerAddressV2_Result
      type: object
      properties:
        ChangeCustomerAddressV2_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeCustomerAddressV2_Result'
          - xml:
              name: ChangeCustomerAddressV2_Result
              attribute: false
              wrapped: false
      xml:
        name: ChangeCustomerAddressV2_Result
        attribute: false
        wrapped: false
    ChangeCustomerAddressV2:
      title: ChangeCustomerAddressV2
      required:
      - ChangeCustomerAddressV2
      type: object
      properties:
        ChangeCustomerAddressV2:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeCustomerAddressV2'
          - {}
    ChangeInvoiceAddress:
      title: ChangeInvoiceAddress
      required:
      - ChangeInvoiceAddress
      type: object
      properties:
        ChangeInvoiceAddress:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeInvoiceAddress'
          - {}
    ChangeInvoiceAddress_Result:
      title: ChangeInvoiceAddress_Result
      required:
      - ChangeInvoiceAddress_Result
      type: object
      properties:
        ChangeInvoiceAddress_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ChangeInvoiceAddress_Result'
          - xml:
              name: ChangeInvoiceAddress_Result
              attribute: false
              wrapped: false
      xml:
        name: ChangeInvoiceAddress_Result
        attribute: false
        wrapped: false
...
    InsertPayment:
      title: InsertPayment
      required:
      - InsertPayment
      type: object
      properties:
        InsertPayment:
          allOf:
          - $ref: '#/components/schemas/tns_InsertPayment'
          - {}
    InsertPayment_Result:
      title: InsertPayment_Result
      required:
      - InsertPayment_Result
      type: object
      properties:
        InsertPayment_Result:
          allOf:
          - $ref: '#/components/schemas/tns_InsertPayment_Result'
          - xml:
              name: InsertPayment_Result
              attribute: false
              wrapped: false
      xml:
        name: InsertPayment_Result
        attribute: false
        wrapped: false
    InsertPaymentV2_Result:
      title: InsertPaymentV2_Result
      required:
      - InsertPaymentV2_Result
      type: object
      properties:
        InsertPaymentV2_Result:
          allOf:
          - $ref: '#/components/schemas/tns_InsertPaymentV2_Result'
          - xml:
              name: InsertPaymentV2_Result
              attribute: false
              wrapped: false
      xml:
        name: InsertPaymentV2_Result
        attribute: false
        wrapped: false
...
    InsertPaymentV2:
      title: InsertPaymentV2
      required:
      - InsertPaymentV2
      type: object
      properties:
        InsertPaymentV2:
          allOf:
          - $ref: '#/components/schemas/tns_InsertPaymentV2'
          - {}
    InsertPaymentV3:
      title: InsertPaymentV3
      required:
      - InsertPaymentV3
      type: object
      properties:
        InsertPaymentV3:
          allOf:
          - $ref: '#/components/schemas/tns_InsertPaymentV3'
          - {}
...
    ModifyInvoice:
      title: ModifyInvoice
      required:
      - ModifyInvoice
      type: object
      properties:
        ModifyInvoice:
          allOf:
          - $ref: '#/components/schemas/tns_ModifyInvoice'
          - {}
    ModifyInvoice_Result:
      title: ModifyInvoice_Result
      required:
      - ModifyInvoice_Result
      type: object
      properties:
        ModifyInvoice_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ModifyInvoice_Result'
          - xml:
              name: ModifyInvoice_Result
              attribute: false
              wrapped: false
      xml:
        name: ModifyInvoice_Result
        attribute: false
        wrapped: false
    ModifyInvoiceV2_Result:
      title: ModifyInvoiceV2_Result
      required:
      - ModifyInvoiceV2_Result
      type: object
      properties:
        ModifyInvoiceV2_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ModifyInvoiceV2_Result'
          - xml:
              name: ModifyInvoiceV2_Result
              attribute: false
              wrapped: false
      xml:
        name: ModifyInvoiceV2_Result
        attribute: false
        wrapped: false
    ModifyInvoiceV3_Result:
      title: ModifyInvoiceV3_Result
      required:
      - ModifyInvoiceV3_Result
      type: object
      properties:
        ModifyInvoiceV3_Result:
          allOf:
          - $ref: '#/components/schemas/tns_ModifyInvoiceV3_Result'
          - xml:
              name: ModifyInvoiceV3_Result
              attribute: false
              wrapped: false
      xml:
        name: ModifyInvoiceV3_Result
        attribute: false
        wrapped: false
...
    ModifyInvoiceV2:
      title: ModifyInvoiceV2
      required:
      - ModifyInvoiceV2
      type: object
      properties:
        ModifyInvoiceV2:
          allOf:
          - $ref: '#/components/schemas/tns_ModifyInvoiceV2'
          - {}
    ModifyInvoiceV3:
      title: ModifyInvoiceV3
      required:
      - ModifyInvoiceV3
      type: object
      properties:
        ModifyInvoiceV3:
          allOf:
          - $ref: '#/components/schemas/tns_ModifyInvoiceV3'
          - {}
...
    ChangeInvoiceAddress_body:
      allOf:
      - $ref: '#/components/schemas/ChangeInvoiceAddress'
    inline_response_200:
      allOf:
      - $ref: '#/components/schemas/ChangeInvoiceAddress_Result'
      - xml:
          name: ChangeInvoiceAddress_Result
          attribute: false
          wrapped: false
    ChangeCustomerAddress_body:
      allOf:
      - $ref: '#/components/schemas/ChangeCustomerAddress'
      - xml:
          name: ChangeCustomerAddress
          attribute: false
          wrapped: false
    inline_response_200_1:
      allOf:
      - $ref: '#/components/schemas/ChangeCustomerAddress_Result'
      - xml:
          name: ChangeCustomerAddress_Result
          attribute: false
          wrapped: false
    ChangeContactInformation_body:
      allOf:
      - $ref: '#/components/schemas/ChangeContactInformation'
      - xml:
          name: ChangeContactInformation
          attribute: false
          wrapped: false
    inline_response_200_2:
      allOf:
      - $ref: '#/components/schemas/ChangeContactInformation_Result'
      - xml:
          name: ChangeContactInformation_Result
          attribute: false
          wrapped: false
    InsertPayment_body:
      allOf:
      - $ref: '#/components/schemas/InsertPayment'
      - xml:
          name: InsertPayment
          attribute: false
          wrapped: false
    inline_response_200_3:
      allOf:
      - $ref: '#/components/schemas/InsertPayment_Result'
      - xml:
          name: InsertPayment_Result
          attribute: false
          wrapped: false
    ModifyInvoice_body:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoice'
      - xml:
          name: ModifyInvoice
          attribute: false
          wrapped: false
    inline_response_200_4:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoice_Result'
      - xml:
          name: ModifyInvoice_Result
          attribute: false
          wrapped: false
    InsertPaymentV2_body:
      allOf:
      - $ref: '#/components/schemas/InsertPaymentV2'
      - xml:
          name: InsertPaymentV2
          attribute: false
          wrapped: false
    inline_response_200_5:
      allOf:
      - $ref: '#/components/schemas/InsertPaymentV2_Result'
      - xml:
          name: InsertPaymentV2_Result
          attribute: false
          wrapped: false
    ChangeCustomerAddressV2_body:
      allOf:
      - $ref: '#/components/schemas/ChangeCustomerAddressV2'
      - xml:
          name: ChangeCustomerAddressV2
          attribute: false
          wrapped: false
    inline_response_200_6:
      allOf:
      - $ref: '#/components/schemas/ChangeCustomerAddressV2_Result'
      - xml:
          name: ChangeCustomerAddressV2_Result
          attribute: false
          wrapped: false
    ModifyInvoiceV2_body:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoiceV2'
      - xml:
          name: ModifyInvoiceV2
          attribute: false
          wrapped: false
    inline_response_200_7:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoiceV2_Result'
      - xml:
          name: ModifyInvoiceV2_Result
          attribute: false
          wrapped: false
    ModifyInvoiceV3_body:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoiceV3'
      - xml:
          name: ModifyInvoiceV3
          attribute: false
          wrapped: false
    inline_response_200_8:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoiceV3_Result'
      - xml:
          name: ModifyInvoiceV3_Result
          attribute: false
          wrapped: false
    InsertPaymentV3_body:
      allOf:
      - $ref: '#/components/schemas/InsertPaymentV3'
      - xml:
          name: InsertPaymentV3
          attribute: false
          wrapped: false
    inline_response_200_9:
      allOf:
      - $ref: '#/components/schemas/InsertPaymentV3_Result'
      - xml:
          name: InsertPaymentV3_Result
          attribute: false
          wrapped: false
    ModifyInvoiceV4_body:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoiceV4'
      - xml:
          name: ModifyInvoiceV4
          attribute: false
          wrapped: false
    inline_response_200_10:
      allOf:
      - $ref: '#/components/schemas/ModifyInvoiceV4_Result'
      - xml:
          name: ModifyInvoiceV4_Result
          attribute: false
          wrapped: false
    ConfirmInvoice_body:
      allOf:
      - $ref: '#/components/schemas/ConfirmInvoice'
      - xml:
          name: ConfirmInvoice
          attribute: false
          wrapped: false
    inline_response_200_11:
      allOf:
      - $ref: '#/components/schemas/ConfirmInvoice_Result'
      - xml:
          name: ConfirmInvoice_Result
          attribute: false
          wrapped: false
    ModifyHotelCreditCardInfo_body:
      allOf:
      - $ref: '#/components/schemas/ModifyHotelCreditCardInfo'
      - xml:
          name: ModifyHotelCreditCardInfo
          attribute: false
          wrapped: false
    inline_response_200_12:
      allOf:
      - $ref: '#/components/schemas/ModifyHotelCreditCardInfo_Result'
      - xml:
          name: ModifyHotelCreditCardInfo_Result
          attribute: false
          wrapped: false
    ModifyHotelBankAccountInfo_body:
      allOf:
      - $ref: '#/components/schemas/ModifyHotelBankAccountInfo'
      - xml:
          name: ModifyHotelBankAccountInfo
          attribute: false
          wrapped: false
    inline_response_200_13:
      allOf:
      - $ref: '#/components/schemas/ModifyHotelBankAccountInfo_Result'
      - xml:
          name: ModifyHotelBankAccountInfo_Result
          attribute: false
          wrapped: false
    InsertPaymentPC_body:
      allOf:
      - $ref: '#/components/schemas/InsertPaymentPC'
      - xml:
          name: InsertPaymentPC
          attribute: false
          wrapped: false
    inline_response_200_14:
      allOf:
      - $ref: '#/components/schemas/InsertPaymentPC_Result'
      - xml:
          name: InsertPaymentPC_Result
          attribute: false
          wrapped: false